+++ CHANGELOG.md
@@ -16,7 +16,11 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 - [SearchOptions] Remove SearchOptions.AttributeSets parameter.
 
+- [SearchEngine] Add RetrieveOptions(attributeSets:) parameter to `select()` functions. Use this to provide attribute set queries.
+
 - [SearchResultMetadata] Add fields for more metadata characteristics, options, social media handles, and contact information.
+
+- [SearchResult] Add ResultChildMetadata type to SearchResultMetadata.children field.
 
 **MapboxCommon**: v24.6.0
 **MapboxCoreSearch**: v2.3.0-rc.6